Tested buffers and fixed scan issues
Implemented and tested input buffers for UART modules.
Apparently sscanf format specifiers need precision values for float and
double values.
NMEA parsing tested and working.

+++ Pinpoint.cydsn/nmea.h
@@ -73,7 +73,7 @@ typedef struct RMC_Str {
    char    lonDir; // E/W
    float   groundSpeed; // In knots
    float   groundCourse; // In degrees
-   uint16_t date;
+   uint32_t date;
    float   magVar; // Degrees
    char    magVarDir; // E/W
    char    mode; // A = autonomous, D = differential, E = estimated
